+++ src/gradient/Gradient.ts
@@ -1,4 +1,3 @@
-import { Color } from '../color/Color';
 import { iMatrix } from '../constants';
 import { parseTransformAttribute } from '../parser/parseTransformAttribute';
 import type { FabricObject } from '../shapes/Object/FabricObject';
@@ -132,11 +131,9 @@ export class Gradient<
    */
   addColorStop(colorStops: Record<string, string>) {
     for (const position in colorStops) {
-      const color = new Color(colorStops[position]);
       this.colorStops.push({
         offset: parseFloat(position),
-        color: color.toRgb(),
-        opacity: color.getAlpha(),
+        color: colorStops[position],
       });
     }
     return this;
@@ -270,15 +267,9 @@ export class Gradient<
       }
     }
 
-    colorStops.forEach(({ color, offset, opacity }) => {
+    colorStops.forEach(({ color, offset }) => {
       markup.push(
-        '<stop ',
-        'offset="',
-        offset * 100 + '%',
-        '" style="stop-color:',
-        color,
-        typeof opacity !== 'undefined' ? ';stop-opacity: ' + opacity : ';',
-        '"/>\n',
+        `<stop offset="${offset * 100}%" style="stop-color:${color};"/>\n`,
       );
     });
 
@@ -303,13 +294,8 @@ export class Gradient<
         ? ctx.createLinearGradient(x1, y1, x2, y2)
         : ctx.createRadialGradient(x1, y1, r1, x2, y2, r2);
 
-    this.colorStops.forEach(({ color, opacity, offset }) => {
-      gradient.addColorStop(
-        offset,
-        typeof opacity !== 'undefined'
-          ? new Color(color).setAlpha(opacity).toRgba()
-          : color,
-      );
+    this.colorStops.forEach(({ color, offset }) => {
+      gradient.addColorStop(offset, color);
     });
 
     return gradient;
@@ -317,7 +303,7 @@ export class Gradient<
 
   static async fromObject(
     options: GradientOptions<'linear'>,
-  ): Promise<Gradient<'radial'>>;
+  ): Promise<Gradient<'linear'>>;
   static async fromObject(
     options: GradientOptions<'radial'>,
   ): Promise<Gradient<'radial'>>;