+++ test/unit/parser.js
@@ -466,10 +466,8 @@
     '</svg>';
 
     fabric.loadSVGFromString(string).then(({ objects }) => {
-      assert.equal(objects[0].fill.colorStops[0].opacity, 0.5);
-      assert.equal(objects[0].fill.colorStops[0].color, 'rgb(255,0,0)');
-      assert.equal(objects[0].fill.colorStops[1].opacity, 0.25);
-      assert.equal(objects[0].fill.colorStops[1].color, 'rgb(0,255,0)');
+      assert.equal(objects[0].fill.colorStops[0].color, 'rgba(255,0,0,0.5)');
+      assert.equal(objects[0].fill.colorStops[1].color, 'rgba(0,255,0,0.25)');
       done();
     });
   });